LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/focus-visible-originating-element-expected.txt



FAIL CSS Test: ::before matches when originating element has :focus-visible selector assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
